added matrices

+++ SVDcpp/runSVD.cpp
@@ -2,25 +2,25 @@
 #include <ctime>
 
 int main (int argc, char **argv){
-  int start_s=clock();
+  double start_s=clock();
   if (argc < 3){
     std::cout<<"SVD needs three input arguments"<<std::endl;
     return 0;
    }
  SVD mySVD = SVD();
- int start_r = clock();
+ double start_r = clock();
  mySVD.ReadFile(argv[1]);
- int stop_r = clock();
+ double stop_r = clock();
 
- int start_cal=clock();
+ double start_cal=clock();
  mySVD.Calc_svd();
- int stop_cal = clock();
+ double stop_cal = clock();
 
- int start_w=clock();
+ double start_w=clock();
  mySVD.WriteFile(argv[2]);
- int stop_w = clock();
+ double stop_w = clock();
  // mySVD.Print();
- int stop_s=clock();
+ double stop_s=clock();
  std::cout << "total time: " << (stop_s-start_s)/double(CLOCKS_PER_SEC)*1000 << std::endl;
  std::cout << "Reading time: " << (stop_r-start_r)/double(CLOCKS_PER_SEC)*1000 << std::endl;
  std::cout << "SVD calculation time: " << (stop_cal-start_cal)/double(CLOCKS_PER_SEC)*1000 << std::endl;
